refactor: extract configuration from constants
- Split config/index.ts into site/domain/navigation configs
- Created config/site.config.ts for site information
- Created config/domain.config.ts for categories and divisions
- Created config/navigation.config.ts for routes
- Updated 21 files with new import statements
- Maintained backward compatibility via re-exports
- No functional changes, pure refactoring

Resolves #4
