+++ src/shared/components/layout/Header.tsx
@@ -8,7 +8,8 @@
 import Link from 'next/link';
 import { useState, useEffect } from 'react';
 import { Menu, X, Leaf } from 'lucide-react';
-import { SITE_CONFIG, ROUTES } from '@/config';
+import { SITE_CONFIG } from '@/config/site.config';
+import { ROUTES } from '@/config/navigation.config';
 import { Navigation } from './Navigation';
 import { MobileMenu } from './MobileMenu';
 import { cn } from '@/shared/utils/cn';
