+++ src/shared/components/layout/Navigation.tsx
@@ -7,7 +7,7 @@
 
 import Link from 'next/link';
 import { usePathname } from 'next/navigation';
-import { ROUTES } from '@/config';
+import { ROUTES } from '@/config/navigation.config';
 import { cn } from '@/shared/utils/cn';
 import { motion } from 'framer-motion';
